Move test suite from "util" to "retry"
